Specify multiple NetworkModels for a Waiter
