+++ src/index.ts
@@ -1,3 +1,3 @@
-export {Waiter} from './waiter'
+export {Waiter, NetworkMap} from './waiter'
 export * from './network'
 export * from './elements'
