+++ src/network.ts
@@ -1,7 +1,6 @@
 
 import {
   NodeId,
-  Action,
   Entry,
   Event,
   EffectAbstract,
@@ -13,21 +12,21 @@ import {
 
 export class NetworkModel {
   // // the total set of agents participating in the network
-  agents: Array<NodeId>
+  nodes: Array<NodeId>
 
   // // a connectivity matrix representing every possible p2p connection
   // // the number represents additional latency to simulate. Negative numbers mean infinite latency.
   // connectionMatrix: Array<Array<number>>
 
-  constructor (agents: Array<NodeId>) {
-    this.agents = agents
+  constructor (nodes: Array<NodeId>) {
+    this.nodes = nodes
   }
 
   /**
    * Given a Cause, determine the Effects.
    * Ideally this will use a very general definition of Action causation,
    * decoupled from any concrete network. This function reifies those causation rules
-   * in the context of this actual network, i.e. determines exactly which agents are involved in a
+   * in the context of this actual network, i.e. determines exactly which nodes are involved in a
    * particular cause-effect pair.
    */
   determineEffects = (o: Observation): Array<EffectConcrete> => {
@@ -56,27 +55,19 @@ export class NetworkModel {
   }
 
   /**
-   * Determine which agents are validators for a particular Entry,
+   * Determine which nodes are validators for a particular Entry,
    * based on the action (which should have a reference to the entry)
    */
   validators = (event: Event): Array<NodeId> => {
     throw "abstract method not implemented"
   }
 
-
-  /**
-   * Not sure what this will be. This is the part that returns promises for each awaited observation
-   * from each agent, along with stats about the time took, timeouts, etc.
-   */
-  resolveObservations = async (observations: Array<Observation>) => {
-
-  }
 }
 
 export class FullSyncNetwork extends NetworkModel {
 
   validators = (entry: Entry): Array<NodeId> => {
     // currently full-sync
-    return this.agents
+    return this.nodes
   }
 }
