+++ test/invariants.ts
@@ -1,0 +1,3 @@
+
+
+// TODO: test that all NodeIDs in a Waiter must be unique
